Add dto validation and custom response types.

+++ TestHub.Application/features/answer/handlers/commands/DeleteAnswerCommandHandler.cs
@@ -1,9 +1,8 @@
 ﻿using Application.features.answer.requests.commands;
-using Application.persistence.contracts;
 
 namespace Application.features.answer.handlers.commands;
 
-public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand, Unit>
+public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand, BaseCommandResponse>
 {
     private readonly IAnswerRepository _repository;
 
@@ -12,11 +11,12 @@ public class DeleteAnswerCommandHandler : IRequestHandler<DeleteAnswerCommand, U
         _repository = repository;
     }
     
-    public async Task<Unit> Handle(DeleteAnswerCommand command, CancellationToken cancellationToken)
+    public async Task<BaseCommandResponse> Handle(DeleteAnswerCommand command, CancellationToken cancellationToken)
     {
         var answer = await _repository.Get(command.Id);
-
+        if (answer == null) return new NotFoundFailedStatusResponse(command.Id);
+        
         await _repository.Delete(answer);
-        return Unit.Value;
+        return new OkSuccessStatusResponse(command.Id);
     }
 }
