+++ TestHub.Application/features/category/handlers/commands/DeleteCategoryCommandHandler.cs
@@ -1,9 +1,8 @@
 ﻿using Application.features.category.requests.commands;
-using Application.persistence.contracts;
 
 namespace Application.features.category.handlers.commands;
 
-public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand, Unit>
+public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand, BaseCommandResponse>
 {
     private readonly ICategoryRepository _repository;
 
@@ -12,11 +11,12 @@ public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryComman
         _repository = repository;
     }
     
-    public async Task<Unit> Handle(DeleteCategoryCommand command, CancellationToken cancellationToken)
+    public async Task<BaseCommandResponse> Handle(DeleteCategoryCommand command, CancellationToken cancellationToken)
     {
         var category = await _repository.Get(command.Id);
+        if (category == null) return new NotFoundFailedStatusResponse(command.Id);
 
         await _repository.Delete(category);
-        return Unit.Value;
+        return new OkSuccessStatusResponse(command.Id);
     }
 }
